ci: use MAGE_OS_CI_TOKEN
When we enabled terraform across the organization, it also enabled
branch protection on the repos. As a result, the user in the CI job was not
allowed to override the branch policies of the repos (min reviewers,
push protections, etc).

We set a organization token called MAGE_OS_CI_TOKEN that is on the
"mage-os-ci" github user https://github.com/mage-os-ci and also set as
an organization secret in MageOS that should allow us to allow this
particular user to overcome branch policies in Github.